nonebot2/requirements.txt
nonebot2==2.0.0b3
nonebot-adapter-onebot==2.1.0
nonebot-plugin-apscheduler==0.1.3
httpx==0.23.0
aiocqhttp==1.4.3
Pillow==9.1.1
fuzzywuzzy==0.18.0
bs4==0.0.1